Big refactor to avoid circular dependency
