+++ src/lib/settings/types/UserSettings.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-namespace */
 import { HexColorString } from 'discord.js';
 
-import { CompostName } from '../../../mahoji/commands/farming';
 import { BitField } from '../../constants';
 import { GearSetup } from '../../gear';
 import { CombatOptionsEnum } from '../../minions/data/combatConstants';
 import { BlowpipeData } from '../../minions/types';
+import { CompostName } from '../../skilling/skills/farming';
 import { BirdhouseData } from '../../skilling/skills/hunter/defaultBirdHouseTrap';
 import { SkillsEnum } from '../../skilling/types';
 import { SlayerTaskUnlocksEnum } from '../../slayer/slayerUnlocks';
