+++ src/lib/skilling/skills/farming/index.ts
@@ -1,4 +1,5 @@
 import { Emoji } from '../../../constants';
+import getOSItem from '../../../util/getOSItem';
 import itemID from '../../../util/itemID';
 import { Plant, SkillsEnum } from '../../types';
 import allotmentPlants from './allotments';
@@ -32,6 +33,22 @@ const femaleFarmerItems: { [key: number]: number } = {
 };
 
 export const allFarmingItems: number[] = [];
+
+export const CompostTiers = [
+	{
+		name: 'compost',
+		item: getOSItem('Compost')
+	},
+	{
+		name: 'supercompost',
+		item: getOSItem('Supercompost')
+	},
+	{
+		name: 'ultracompost',
+		item: getOSItem('Ultracompost')
+	}
+] as const;
+export type CompostName = typeof CompostTiers[number]['name'];
 
 for (const plant of plants) {
 	if (plant.outputCrop) allFarmingItems.push(plant.outputCrop);
