+++ src/mahoji/commands/farming.ts
@@ -1,155 +1,20 @@
-import { time } from '@discordjs/builders';
-import { User } from '@prisma/client';
-import { Time } from 'e';
 import { APIUser, ApplicationCommandOptionType, CommandRunOptions } from 'mahoji';
 
 import { client } from '../..';
-import { Emoji } from '../../lib/constants';
 import TitheFarmBuyables from '../../lib/data/buyables/titheFarmBuyables';
 import { superCompostables } from '../../lib/data/filterables';
-import { defaultPatches } from '../../lib/minions/farming';
-import { ContractOption, ContractOptions, IPatchData, IPatchDataDetailed } from '../../lib/minions/farming/types';
+import { ContractOption, ContractOptions } from '../../lib/minions/farming/types';
 import { autoFarm } from '../../lib/minions/functions/autoFarm';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
-import Farming from '../../lib/skilling/skills/farming';
-import { assert, formatDuration, stringMatches, toTitleCase } from '../../lib/util';
-import getOSItem from '../../lib/util/getOSItem';
+import { getFarmingInfo } from '../../lib/skilling/functions/getFarmingInfo';
+import Farming, { CompostName, CompostTiers } from '../../lib/skilling/skills/farming';
+import { stringMatches } from '../../lib/util';
+import { farmingPatchNames, userGrowingProgressStr } from '../../lib/util/farmingHelpers';
 import { compostBinCommand, farmingPlantCommand, harvestCommand } from '../lib/abstracted_commands/farmingCommand';
 import { faringContractCommand } from '../lib/abstracted_commands/farmingContractCommand';
 import { titheFarmCommand, titheFarmShopCommand } from '../lib/abstracted_commands/titheFarmCommand';
 import { OSBMahojiCommand } from '../lib/util';
 import { getSkillsOfMahojiUser, mahojiUserSettingsUpdate, mahojiUsersSettingsFetch } from '../mahojiSettings';
-
-export const farmingPatchNames = [
-	'herb',
-	'fruit_tree',
-	'tree',
-	'allotment',
-	'hops',
-	'cactus',
-	'bush',
-	'spirit',
-	'hardwood',
-	'seaweed',
-	'vine',
-	'calquat',
-	'redwood',
-	'crystal',
-	'celastrus',
-	'hespori',
-	'flower',
-	'mushroom',
-	'belladonna'
-] as const;
-
-export type FarmingPatchName = typeof farmingPatchNames[number];
-
-export function isPatchName(name: string): name is FarmingPatchName {
-	return farmingPatchNames.includes(name as FarmingPatchName);
-}
-
-const farmingKeys: (keyof User)[] = farmingPatchNames.map(i => `farmingPatches_${i}` as const);
-
-export function getFarmingKeyFromName(name: FarmingPatchName): keyof User {
-	return `farmingPatches_${name}`;
-}
-
-export function findPlant(lastPlanted: IPatchData['lastPlanted']) {
-	if (!lastPlanted) return null;
-	const plant = Farming.Plants.find(
-		plants => stringMatches(plants.name, lastPlanted) || plants.aliases.some(a => stringMatches(a, lastPlanted))
-	);
-	if (!plant) throw new Error(`No plant found for ${lastPlanted}.`);
-	return plant;
-}
-
-export function userGrowingProgressStr(patchesDetailed: IPatchDataDetailed[]) {
-	let str = '';
-	for (const patch of patchesDetailed.filter(i => i.ready === true)) {
-		str += `${Emoji.Tick} **${patch.friendlyName}**: ${patch.lastQuantity} ${patch.lastPlanted} is ready to be harvested!\n`;
-	}
-	for (const patch of patchesDetailed.filter(i => i.ready === false)) {
-		str += `${Emoji.Stopwatch} **${patch.friendlyName}**: ${patch.lastQuantity} ${
-			patch.lastPlanted
-		} ready at ${time(patch.readyAt!, 'T')} (${time(patch.readyAt!, 'R')})\n`;
-	}
-	const notReady = patchesDetailed.filter(i => i.ready === null);
-	str += `${Emoji.RedX} **Nothing planted:** ${notReady.map(i => i.friendlyName).join(', ')}.`;
-	return str;
-}
-
-export async function getFarmingInfo(userID: bigint | string) {
-	let keys: Partial<Record<keyof User, true>> = {};
-	for (const key of farmingKeys) keys[key] = true;
-	const userData = await mahojiUsersSettingsFetch(userID, keys);
-
-	const patches: Record<FarmingPatchName, IPatchData> = {} as Record<FarmingPatchName, IPatchData>;
-	const patchesDetailed: IPatchDataDetailed[] = [];
-
-	const now = Date.now();
-
-	for (const key of farmingKeys) {
-		const patch: IPatchData = (userData[key] as IPatchData | null) ?? defaultPatches;
-		const patchName: FarmingPatchName = key.replace('farmingPatches_', '') as FarmingPatchName;
-		assert(farmingPatchNames.includes(patchName));
-		patches[patchName] = patch;
-
-		const plant = findPlant(patch.lastPlanted);
-		const difference = now - patch.plantTime;
-
-		const ready = plant ? difference > plant.growthTime * Time.Minute : null;
-		const readyAt = plant ? new Date(patch.plantTime + plant.growthTime * Time.Minute) : null;
-		const readyIn = readyAt ? readyAt.getTime() - now : null;
-
-		if (ready) {
-			assert(readyAt !== null, "readyAt shouldn't be null if ready");
-			assert(
-				readyIn !== null && readyIn <= 0,
-				`${patchName} readyIn should be less than 0 ready, received ${readyIn} ${formatDuration(readyIn!)}`
-			);
-		}
-		if (ready === false) {
-			assert(readyAt !== null, `readyAt should have value if growing, received ${readyAt}`);
-			assert(readyAt !== null, `${patchName} readyIn should have value if growing, received ${readyAt}`);
-		}
-		if (ready === null) {
-			assert(readyAt === null, 'No readyAt if null');
-			assert(readyIn === null, 'No readyIn if null');
-		}
-
-		patchesDetailed.push({
-			...patch,
-			ready,
-			readyIn,
-			patchName,
-			readyAt,
-			friendlyName: toTitleCase(patchName.replace('_', ' ')),
-			plant
-		});
-	}
-
-	return {
-		patches,
-		patchesDetailed
-	};
-}
-
-export const CompostTiers = [
-	{
-		name: 'compost',
-		item: getOSItem('Compost')
-	},
-	{
-		name: 'supercompost',
-		item: getOSItem('Supercompost')
-	},
-	{
-		name: 'ultracompost',
-		item: getOSItem('Ultracompost')
-	}
-] as const;
-
-export type CompostName = typeof CompostTiers[number]['name'];
 
 export const farmingCommand: OSBMahojiCommand = {
 	name: 'farming',
